Create README.md
# MGK_UI_BUILDER
Adding Dynamic UI Generator and Config Builder for Tkinter

The idea is to create a simple app that allows users to dynamically create a UI configuration for a Tkinter application. The interface_builder.py file contains a DynamicUIGenerator class that takes a UI configuration and dynamically generates a Tkinter user interface based on it.


Main Files Walkthrough
config_builder.py: This file introduces a UIConfigBuilder class that allows users to dynamically create a UI configuration for a Tkinter application. It also includes a ElementPopupFactory class that creates popups for different UI elements, and an ElementPopup class that serves as a base class for specific element popups like LabelPopup.

interface_builder.py: This file introduces a DynamicUIGenerator class that takes a UI configuration and dynamically generates a Tkinter user interface based on it. It supports various types of widgets such as labels, buttons, entries, checkboxes, frames, menus, and comboboxes. The file also includes sample functions for button click, menu click, and combobox select events.

This is just the first scratch...